Refactor deploying mocks into `unit` script
